requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
numpy==1.24.4
transformers==4.37.0
faiss-cpu==1.7.4
PyPDF2
boto3
